services/order-service/requirements.txt
django
djangorestframework
pika
psycopg2-binary
dj-database-url
